feat: add preview and report email sending
Related with the preview, when the report is generated, a button is displayed in the screen, and the user can download it to see if the report is ok. Once he/she decides that the report is ok, he/she can click ont a button that will send it to the email of the selected donor.

Inorder to make the report attachment possible, it was necessary to modify the email sending function of request_info_update.py